# web-poll-swagger/api-gateway/api/docs/swagger.yaml
definitions:
...
  genprotos.Feedback:
    properties:
      from:
        type: integer
      text:
        type: string
      to:
        type: integer
    type: object
...
  genprotos.Option:
    properties:
      ball:
        type: integer
      variant:
        type: string
    type: object
...
  genprotos.PollGetByIDRes:
    properties:
      feedback:
        items:
          $ref: '#/definitions/genprotos.Feedback'
        type: array
      id:
        type: string
      options:
        items:
          $ref: '#/definitions/genprotos.Option'
        type: array
      poll_num:
        type: integer
      title:
        type: string
    type: object
...
  genprotos.Question:
    properties:
      content:
        type: string
      id:
        type: string
      num:
        type: integer
      poll_id:
        type: string
    type: object
  genprotos.QuestionCreateReq:
    properties:
      content:
        type: string
      poll_id:
        type: string
    type: object
  genprotos.QuestionGetAllRes:
    properties:
      poll:
        $ref: '#/definitions/genprotos.PollGetByIDRes'
      question:
        items:
          $ref: '#/definitions/genprotos.Question'
        type: array
    type: object
  genprotos.QuestionGetByIDRes:
    properties:
      content:
        type: string
      id:
        type: string
      poll_id:
        type: string
    type: object
  genprotos.QuestionUpdateReq:
    properties:
      content:
        type: string
      id:
        type: string
    type: object
...
info:
  contact: {}
paths:
...
  /question:
    post:
      consumes:
      - application/json
      description: Create a new question with the given details
      parameters:
      - description: Question creation request
        in: body
        name: question
        required: true
        schema:
          $ref: '#/definitions/genprotos.QuestionCreateReq'
      produces:
      - application/json
      responses:
        "200":
          description: Question created successfully
          schema:
            type: string
        "400":
          description: Invalid request payload
          schema:
            type: string
        "500":
          description: Server error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Create a new question
      tags:
      - question
    put:
      consumes:
      - application/json
      description: Update an existing question with the given details
      parameters:
      - description: Question update request
        in: body
        name: question
        required: true
        schema:
          $ref: '#/definitions/genprotos.QuestionUpdateReq'
      produces:
      - application/json
      responses:
        "200":
          description: Question updated successfully
          schema:
            type: string
        "400":
          description: Invalid request payload
          schema:
            type: string
        "500":
          description: Server error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Update an existing question
      tags:
      - question
  /question/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a question by its ID
      parameters:
      - description: Question ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Question deleted successfully
          schema:
            type: string
        "500":
          description: Server error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Delete a question
      tags:
      - question
    get:
      consumes:
      - application/json
      description: Get a question by its ID
      parameters:
      - description: Question ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/genprotos.QuestionGetByIDRes'
        "500":
          description: Server error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Get a question by ID
      tags:
      - question
  /questions/{poll_id}:
    get:
      consumes:
      - application/json
      description: Get all questions with pagination
      parameters:
      - description: Poll ID
        in: path
        name: poll_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/genprotos.QuestionGetAllRes'
        "400":
          description: Invalid request payload
          schema:
            type: string
        "500":
          description: Server error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Get all questions
      tags:
      - question
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"